Curated batch_86_update
Curated batch_86_update

For Mucopolysaccharidosis, MPS-I-Laronidase see this bit of the SOP: Paths should be logically consistent “If all of the relationships in a path are relationships of influence (+1 for increases or -1 for decreases), the product of their directions (e.g. the overall direction of influence of the path) should be negative”. In this graph’s direction of influence of the path is positive.

Lakshman: Yes understood, have changed and simplified the path

Age related macular degeneration-lutein. Part of the path which includes Macula and retina has positive direction of influence. Also not sure how a blue light photoreceptor activity prevents the macular degeneration?


Lakshman: blue light photoreceptor activity was performed by lutein in natural conditions. But I agree that it may not be relevant with respect to AMD.

Urge incontinence of urine- trospium: please double check if M1 receptor is involved in bladder smooth muscle contraction

Lakshman: Checked the literature, M3 receptor is the major one involved in bladder smooth muscle contraction, changed it .

5) Wilson's disease-: One path is in this graph has positive direction of influence.

`Cooper homeostasis` is not a manifestation of Wilson’s disease. 

Lakshman: Changed it to another predicate "prevents"

@jaganlakshman Please correct these graphs, and review other graphs in this pull request and make another pull request. Thank you!

Overall: Changed the format for the "comment" section, modified the reference section as per the latest guidelines. Added path ID to all.