commit 86_Updated2
commit 86_Updated2

@jaganlakshman thanks for the corrections. I would like to add a few things:
1)The entry term GO:0060073 is micturition not urination.

Lakshman: Changed to micturition

2) Graph Wilson's disease- penicillamine: Copper regulates regulation of excretion – does not sound right to me.

Lakshman: Removed the "regulation of excretion" 

3) Postherpetic neuralgia- gabapentin enacarbil and Restless legs -gabapentin enacarbil: Gabapentin enacarbil is a gabapentin prodrug, please use the format for the prodrugs that is in the curation guide. In this case : Gabapentin enacarbil has metabolite gabapentin. I’m not sure if the statement: presynapse negatively regulates calcium channel activity and calcium-release channel activity, is correct. The Voltage-dependent calcium channel subunit alpha-2/delta-2 seem to interact with: Neurexins, Thrombospondins and NMDA receptors so predicate decreases molecular interaction does not seem to be correct for this context.

Lakshman: Used the prodrug format i.e., added has metabolite. Since the activity of the  gabapentin enacarbil was same as gabapentin, I have in fact used previous gabapentin annotation from the DrugDB and did not modify any (to keep the consistency). Now have changed the directionality of the predicate kept as regulate instead of increase or decrease statement after the presynapse statement. As per the literature available gabapentin definitely modify calcium channel activity for sure, may be the directionality was not proper I guess. I am not sure who did actual gabapentin curation, may be that needs to be changed as well.

4) In your comments for the PR I see: "aded path ID to all. Since I want to delete all my previous submissions, not adding path IDs." - Just to let you know that when I reviewed this PR I did not see any paths IDs added.

Lakshman: Sorry, this comment I kept before requesting for the deletion of my entire list. Now, since it was deleted and I hope the path ID was not required.